storage: Move NFS management into new details page
This gets rid of the confusing "arm" button.

Fixes #8561
Fixes #8565
