+++ pkg/storaged/storage.css
@@ -358,6 +358,10 @@ div.progress {
     width: 100%;
 }
 
+#detail-header div.progress {
+    min-width: 250px;
+}
+
 td.usage-text {
     text-align: right;
 }
